feat: Extract lifecycle event handling and simplify state management (launcher-apps)
This commit refactors the `LifecycleEffect` composable into a new `SyncDataEffect`
and extracts the `ManagedProfileResult` and `statusBarNotifications` state
management into separate composables.

The `SyncDataEffect` now solely focuses on registering and unregistering
`LauncherApps.Callback` and managing `AppWidgetHost` listening. The logic for
managed profiles and status bar notifications has been moved out to
`rememberManagedProfileResult` and `rememberStatusBarNotifications` respectively.

This change improves the modularity and readability of the code, making it
easier to manage and test individual components.

Key changes:
- **Extracted `SyncDataEffect`:** Encapsulates lifecycle-related data synchronization
  logic, including `LauncherApps` callbacks and `AppWidgetHost` management.
- **Extracted `rememberManagedProfileResult`:** Manages the state and updates for
  managed profile information via a `BroadcastReceiver`.
- **Extracted `rememberStatusBarNotifications`:** Manages the state and updates for
  status bar notifications by binding to `EblanNotificationListenerService`.
- **Removed `LifecycleEffect`:** Replaced by the new, more focused composables.
- **Updated `PagerScreen` and `HomeScreen`:** These screens now utilize the
  newly extracted composables for managing their respective states and effects.
- **Simplified `PagerScreenState`:** Removed direct state variables for
  `managedProfileResult` and `statusBarNotifications` as they are now managed
  externally.
- **`OverlayImage` composable moved:** Moved `OverlayImage` into `PagerScreen.kt`
  for better organization.
